server/requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
pymysql==1.0.2
